Move stack tests
