+++ python/cudf/cudf/tests/reshape/test_melt.py
@@ -1,6 +1,5 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
 
-
 import numpy as np
 import pandas as pd
 import pytest
